# Improve find and grep performance