--- a/find.go
+++ b/find.go
@@ -64,13 +64,13 @@ func (f find) findFile(root string, regexp *regexp.Regexp) {
 	}
 
 	followed := f.opts.SearchOption.Follow
-	concurrentWalk(root, ignores, followed, func(path string, info fileInfo, depth int, ignores ignoreMatchers) (ignoreMatchers, error) {
+	concurrentWalk(root, ignores, followed, func(path string, info *fileInfo, depth int, ignores ignoreMatchers) (ignoreMatchers, error) {
 		if info.isDir(followed) {
 			if depth > f.opts.SearchOption.Depth+1 {
 				return ignores, filepath.SkipDir
 			}
 
-			if !f.opts.SearchOption.Hidden && isHidden(info.Name()) {
+			if !f.opts.SearchOption.Hidden && isHidden(info.name) {
 				return ignores, filepath.SkipDir
 			}
 
@@ -91,7 +91,7 @@ func (f find) findFile(root string, regexp *regexp.Regexp) {
 			return ignores, nil
 		}
 
-		if !f.opts.SearchOption.Hidden && isHidden(info.Name()) {
+		if !f.opts.SearchOption.Hidden && isHidden(info.name) {
 			return ignores, filepath.SkipDir
 		}