--- a/readdir_dirent_ino.go
+++ b/readdir_dirent_ino.go
@@ -1,0 +1,9 @@
+// +build linux,!appengine darwin
+
+package the_platinum_searcher
+
+import "syscall"
+
+func direntInode(dirent *syscall.Dirent) uint64 {
+	return uint64(dirent.Ino)
+}