--- a/walk.go
+++ b/walk.go
@@ -1,13 +1,12 @@
 package the_platinum_searcher
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"sync"
 )
 
-type walkFunc func(path string, info fileInfo, depth int, ignores ignoreMatchers) (ignoreMatchers, error)
+type walkFunc func(path string, info *fileInfo, depth int, ignores ignoreMatchers) (ignoreMatchers, error)
 
 func concurrentWalk(root string, ignores ignoreMatchers, followed bool, walkFn walkFunc) error {
 	info, err := os.Lstat(root)
@@ -15,13 +14,13 @@ func concurrentWalk(root string, ignores ignoreMatchers, followed bool, walkFn w
 		return err
 	}
 	sem := make(chan struct{}, 16)
-	return walk(root, newFileInfo(root, info), 1, ignores, followed, walkFn, sem)
+	return walk(root, newFileInfo(root, info.Name(), info.Mode()), 1, ignores, followed, walkFn, sem)
 }
 
-func walk(path string, info fileInfo, depth int, parentIgnores ignoreMatchers, followed bool, walkFn walkFunc, sem chan struct{}) error {
+func walk(path string, info *fileInfo, depth int, parentIgnores ignoreMatchers, followed bool, walkFn walkFunc, sem chan struct{}) error {
 	ignores, walkError := walkFn(path, info, depth, parentIgnores)
 	if walkError != nil {
-		if info.IsDir() && walkError == filepath.SkipDir {
+		if info.isDir(false) && walkError == filepath.SkipDir {
 			return nil
 		}
 		return walkError
@@ -31,7 +30,7 @@ func walk(path string, info fileInfo, depth int, parentIgnores ignoreMatchers, f
 		return nil
 	}
 
-	files, err := ioutil.ReadDir(path)
+	files, err := readDir(path)
 	if err != nil {
 		return err
 	}
@@ -39,17 +38,18 @@ func walk(path string, info fileInfo, depth int, parentIgnores ignoreMatchers, f
 	depth++
 	wg := &sync.WaitGroup{}
 	for _, file := range files {
-		f := newFileInfo(path, file)
 		select {
 		case sem <- struct{}{}:
 			wg.Add(1)
-			go func(path string, file fileInfo, depth int, ignores ignoreMatchers, wg *sync.WaitGroup) {
-				defer wg.Done()
-				defer func() { <-sem }()
+			go func(path string, file *fileInfo, depth int, ignores ignoreMatchers, wg *sync.WaitGroup) {
+				defer func() {
+					wg.Done()
+					<-sem
+				}()
 				walk(path, file, depth, ignores, followed, walkFn, sem)
-			}(filepath.Join(path, file.Name()), f, depth, ignores, wg)
+			}(path+string(os.PathSeparator)+file.name, file, depth, ignores, wg)
 		default:
-			walk(filepath.Join(path, file.Name()), f, depth, ignores, followed, walkFn, sem)
+			walk(path+string(os.PathSeparator)+file.name, file, depth, ignores, followed, walkFn, sem)
 		}
 	}
 	wg.Wait()