1. Use a faster `readdir` implement from [this](https://github.com/golang/tools/commit/edf8e6fef861d8c5ecc9e394b0146d6ebba57795).
2. Buffer output print.


Here is the benchmark result:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkPT/findOnly-4     171367748     31854579      -81.41%
BenchmarkPT/normal-4       302596661     280387391     -7.34%

benchmark                  old allocs     new allocs     delta
BenchmarkPT/findOnly-4     254596         172796         -32.13%
BenchmarkPT/normal-4       635247         552380         -13.04%

benchmark                  old bytes     new bytes     delta
BenchmarkPT/findOnly-4     45581341      9212957       -79.79%
BenchmarkPT/normal-4       366203739     329435411     -10.04%
```

Should fix issue #178 .