@monochromegane 2018-07-07
I confirmed the improvement in performance.
Thanks a lot !!